+++ lib/widgets/home_page/stats/info_tab.dart
